Call after function with number rendered after remove
Fix so that when clicking to remove a rendered element or entity
the correct number of entities rendered is passed to the _after()
function, which then passes the correct number to the defined (if any)
after() function. Without this change the _after() function is always
passed -1 and therefore after() is always passed -1.