Align Docker documentation with current pyFF endpoint behavior
Patch to align the Docker documentation with the current pyFF endpoint
behavior. The MDQ_URL needs to end with entities\ (note the slash). If
it does not then pyFF currently will return metadata for all entities
instead of the metadata for the single entity the user wants to use for
authentication. I also added a slash for SEARCH_URL for consistency
(tested and it works).